refactor(components): move FORM_ITEM_INJECTION_KEY to a different module
By exporting the  from the  file, we can remove the need of having
multiple script tags inside the component.
This change helps with the linters and formatters, expecially with  defaults.
