+++ apps/www/src/lib/registry/new-york/ui/form/injectionKeys.ts
@@ -1,0 +1,4 @@
+import type { InjectionKey } from 'vue'
+
+export const FORM_ITEM_INJECTION_KEY
+  = Symbol() as InjectionKey<string>
